Move add punishment call
