+++ frontend/src/components/leaderboard/createPunishmentTableRow/CreatePunishmentTableRow.tsx
@@ -6,48 +6,31 @@ import { useState } from "react";
 import { Leaderboard } from "../../../helpers/types";
 
 interface CreatePunishmentTableRowProps {
-  groupId: number;
-  userId: number;
+  newPunishment: {
+    punishment_type_id: number;
+    reason: string;
+    reason_hidden: boolean;
+    amount: number;
+  };
+  setNewPunishment: React.Dispatch<
+    React.SetStateAction<{
+      punishment_type_id: number;
+      reason: string;
+      reason_hidden: boolean;
+      amount: number;
+    }>
+  >;
   data: Leaderboard;
 }
 
 export const CreatePunishmentTableRow = ({
-  groupId,
-  userId,
+  newPunishment,
+  setNewPunishment,
   data,
-}: CreatePunishmentTableRowProps) => {
-  const ADD_PUNISHMENT_URL = getAddPunishmentUrl(groupId, userId);
-  const [newPunishment, setNewPunishment] = useState({
-    punishment_type_id: 1,
-    reason: "",
-    reason_hidden: false,
-    amount: 0,
-  });
-
-  const createPunishmentCall = async () => {
-    const res: AxiosResponse<string> = await axios.post(ADD_PUNISHMENT_URL, [
-      newPunishment,
-    ]);
-    return res.data;
-  };
-
-  const { mutate } = useMutation(createPunishmentCall, {
-    onSuccess: () => {
-      console.log("Todo: Handle success");
-    },
-    onError: () => {
-      console.log("Todo: Handle error");
-    },
-  });
-
-  console.log(newPunishment);
-
-  return (
-    <InputForm
-      newPunishment={newPunishment}
-      setNewPunishment={setNewPunishment}
-      submitClickHandler={mutate}
-      data={data}
-    />
-  );
-};
+}: CreatePunishmentTableRowProps) => (
+  <InputForm
+    newPunishment={newPunishment}
+    setNewPunishment={setNewPunishment}
+    data={data}
+  />
+);
